python_workers/requirements.txt

appium==2.0.0
selenium==4.11.0
playwright==1.36.0
supabase==2.3.1
pytest==7.4.0
pytest-asyncio==0.21.0
requests==2.31.0
python-dotenv==1.0.0
psycopg2==2.9.6
pymongo==4.4.1
numpy==1.25.2
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
keras==2.13.1
flask==2.3.2
fastapi==0.100.1
uvicorn==0.23.1
kubernetes==26.1.0
docker==6.1.3
boto3==1.28.57
google-cloud-functions==1.11.0
google-generativeai